Add telegraf-bin-1.20.3.
Aside from Gentoo's go build system being slightly painful to figure out, there
appears to be a bug that prevents packages with a very large number of files
listed in SRC_URI from working. The 1.18.x branch only had 1200 or so, which was
still okay, but the 1.20.x line has ballooned to more than 3000, which comes up
against some limits. A -bin package, therefore, was much easier to get going.

Signed-off-by: Jason Lynch <jason@calindora.com>